requirements.txt
Flask==2.3.2
gunicorn==21.2.0
redis==4.5.5
requests==2.31.0
opencv-python-headless==4.8.0.74
stripe==7.9.0
openai-whisper==20230314
twilio==8.10.3
setuptools==68.2.2
wheel==0.41.2
